Switch to use a "bootstrap" pattern in Visual Editor.
